refactor: existsByIsbnOrTitle -> existsByIsbn #28
