+++ src/main/java/com/readup/server/book/domain/repository/BookRepository.java
@@ -3,13 +3,13 @@ package com.readup.server.book.domain.repository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import com.readup.server.book.application.dto.GetBookResponse;
 import com.readup.server.book.domain.Book;
-import com.readup.server.book.presentation.dto.GetBookResponse;
 
 public interface BookRepository {
 	Book save(Book newBook);
 
-	boolean existsByIsbnOrTitle(String isbn, String title);
+	boolean existsByIsbn(String isbn);
 
 	Book getBookById(Long id);
